raspberry-pi-iot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
opencv-python==4.8.1.78
numpy==1.24.3
smbus2==0.4.3
RPi.GPIO==0.7.1
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
asyncio-mqtt==0.16.1
Pillow==10.1.0
